+++ CHANGELOG.md
@@ -48,6 +48,9 @@ finality provider distribution info
 - [#1831](https://github.com/babylonlabs-io/babylon/pull/1831) Removed unused parameter in `DeriveTaprootPkScript`
 - [#1832](https://github.com/babylonlabs-io/babylon/pull/1832) Add check for multiple txOut with same pk and value
 in `GetOutputIdxInBTCTx`
+- [#1822](https://github.com/babylonlabs-io/babylon/pull/1822) Support multisig for btc staker in btcstaking library
+- [#1824](https://github.com/babylonlabs-io/babylon/pull/1824) Support multisig for btc staker in babylon genesis
+- [#1827](https://github.com/babylonlabs-io/babylon/pull/1827) Upgrade handler to support multisig
 
 ### Bug fixes